api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
dependency-injector = "^4.41"
asyncpg = "^0.29"
fastapi = "^0.110"
sqlalchemy = "^2.0.28"
hypercorn = {version="^0.16", extras=["uvloop"]}
passlib = "^1.7.4"
python-jose = "^3.3.0"
psycopg-binary = "^3.1.18"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"
pytest = "^8.0.1"
pydantic = {version="^2.7.3"}
email-validator = "^2.1"
pydantic-extra-types = "^2.8.0"
phonenumbers = "^8.13.38"
httpx = "^0.27"
sqlmodel = "^0.0.18"
bcrypt = "^4.1.2"
aioboto3 = "^12.3.0"
pre-commit = "^3.7.0"
aio-pika = "^9.4.1"
motor = "^3.4.0"
redis = "^5.0.0"
starlette-admin = "^0.14"
sqlalchemy-file = "^0.6.0"
pillow = "^10.3.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'